refactor(ext/websocket): use fastwebsockets client (#18725)
